// src/generated/modal_proto/task_command_router.ts
export enum TaskExecStderrConfig {
  /** TASK_EXEC_STDERR_CONFIG_DEVNULL - The output will be discarded. */
  TASK_EXEC_STDERR_CONFIG_DEVNULL = 0,
  /** TASK_EXEC_STDERR_CONFIG_PIPE - The output will be streamed to the client. */
  TASK_EXEC_STDERR_CONFIG_PIPE = 1,
  /**
   * TASK_EXEC_STDERR_CONFIG_STDOUT - A special value that can be used to indicate that the stderr stream should
   * be merged with the stdout stream.
   */
  TASK_EXEC_STDERR_CONFIG_STDOUT = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum TaskExecStdoutConfig {
  /** TASK_EXEC_STDOUT_CONFIG_DEVNULL - The output will be discarded. */
  TASK_EXEC_STDOUT_CONFIG_DEVNULL = 0,
  /** TASK_EXEC_STDOUT_CONFIG_PIPE - The output will be streamed to the client. */
  TASK_EXEC_STDOUT_CONFIG_PIPE = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface TaskExecStartRequest {
  /** The ID of the task to execute the command in. */
  taskId: string;
  /**
   * Execution ID. This ID will be used to identify the execution for other
   * requests and ensure exec commands are idempotent.
   */
  execId: string;
  /** Command arguments to execute. */
  commandArgs: string[];
  /** Configures how the stdout of the command will be handled. */
  stdoutConfig: TaskExecStdoutConfig;
  /** Configures how the stderr of the command will be handled. */
  stderrConfig: TaskExecStderrConfig;
  /**
   * Timeout in seconds for the exec'd command to exit. If the command does not
   * exit within this duration, the command will be killed. This is NOT the
   * timeout for the ExecStartRequest RPC to complete.
   */
  timeoutSecs?:
    | number
    | undefined;
  /** Working directory for the command. */
  workdir?:
    | string
    | undefined;
  /** Secret IDs to mount into the task. */
  secretIds: string[];
  /** PTY info for the command. */
  ptyInfo?:
    | PTYInfo
    | undefined;
  /**
   * Enable debugging capabilities on the container runtime. Used only for
   * internal debugging.
   */
  runtimeDebug: boolean;
  /** Fully qualified target container ID. Empty targets the main container. */
  containerId: string;
  /** Environment variables to set directly for the exec'd command. */
  env: { [key: string]: string };
}